postgres amd memory amd file
