+++ internal/store/memory/memory.go
@@ -1,6 +1,7 @@
 package memory
 
 import (
+	"github.com/evgensr/practicum1/internal/line"
 	"sync"
 )
 
@@ -9,12 +10,7 @@ type Box struct {
 	Items           []Line
 	fileStoragePath string
 }
-type Line struct {
-	User   string `json:"user,omitempty"`
-	Url    string `json:"original_url"`
-	Short  string `json:"short_url"`
-	Status int    `json:"status"`
-}
+type Line = line.Line
 
 // RWMap структура Mutex
 type RWMap struct {
